Include Ian's patch (#669) to see if the build works.
