+++ appveyor.yml
@@ -25,12 +25,6 @@ build_script:
   - conda install conda-build
   - ps: if(-not $env:APPVEYOR_PULL_REQUEST_NUMBER) { conda install anaconda-client }
   - conda build conda.recipe --channel dynd/channel/dev
-  - conda install numpy libdynd --channel dynd/channel/dev
-  - conda install numba
-  - ps: conda install (conda build --output conda.recipe)
-  - cd ..
-  - python -c "import dynd; import sys; dynd.test(verbosity=2, exit=True)"
-  - cd dynd-python
 
 on_success:
   # Redirect output from stderr to stdout to avoid having the command for uploading
