+++ conda.recipe/meta.yaml
@@ -26,9 +26,11 @@ requirements:
     - libdynd
 
 # Test separately to avoid unsatisfiable package dependencies bug on Win32.
-#test:
-#  commands:
-#    - python -c "import dynd; dynd.test(verbosity=2, exit=True)"
+test:
+  requires:
+    - numba
+  commands:
+    - python -c "import dynd; dynd.test(verbosity=2, exit=True)"
 
 about:
   home: http://libdynd.org
