+++ dynd/include/array_from_py.hpp
@@ -29,11 +29,7 @@ namespace pydynd {
  * \param always_copy If this is set to true, a new copy is always
  *                    created.
  */
-PYDYND_API dynd::nd::array array_from_py(PyObject *obj, uint32_t access_flags,
-                                         bool always_copy);
-
-PYDYND_API dynd::ndt::type xtype_for_prefix(PyObject *obj);
-
+PYDYND_API dynd::nd::array array_from_py(PyObject *obj, uint32_t access_flags, bool always_copy);
 
 void init_array_from_py();
