+++ dynd/nd/array.pxd
@@ -11,12 +11,9 @@ cdef api class array(object)[object dynd_nd_array_pywrapper,
 cdef _array as_cpp_array(object obj) except *
 cpdef array asarray(object obj)
 
-cdef api _array dynd_nd_array_to_cpp(array) except *
-cdef api _array *dynd_nd_array_to_ptr(array) except *
+cdef api _array dynd_nd_array_to_cpp(array) nogil except *
+cdef api _array *dynd_nd_array_to_ptr(array) nogil except *
 cdef api array dynd_nd_array_from_cpp(_array)
 
 cdef _callable _functional_apply(_type t, object o) except *
 cdef void _registry_assign_init() except *
-cdef object _numpy_dtype_from__type(const _type &tp)
-cdef bint _is_numpy_dtype(PyObject *o) except *
-cdef _type _xtype_for_prefix(object o) except *
